+++ src/lib/structs.py
@@ -19,6 +19,7 @@ class SParamFile:
         self.tag = tag
 
         self._nw = None
+        self._error = None
 
         self.name = name if name is not None else self.filename
         self.short_name = short_name if short_name is not None else os.path.splitext(self.filename)[0]
@@ -32,11 +33,17 @@ class SParamFile:
                 self._nw = skrf.network.Network(self.file_path)
             except Exception as ex:
                 logging.exception(f'Unable to load network from "{self.file_path}" ({ex})')
+                self._error = str(ex)
+                self._nw = None
         return self._nw
     
 
     def loaded(self) -> bool:
         return self._nw is not None
+    
+
+    def error(self) -> bool:
+        return self._error is not None